requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.0
flake8==6.1.0
pytest==7.4.2
requests==2.31.0
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg-pool==3.2.0